requirements.txt
annotated-types==0.7.0
anyio==4.10.0
awscli==1.42.18
bcrypt==3.2.2
boto3==1.40.17
botocore==1.40.18
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
colorama==0.4.6
docutils==0.19
ecdsa==0.19.1
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
